requirements.txt
Flask==3.1.0
Flask-Caching==2.3.0
flasgger==0.9.7.1
redis==5.2.1
requests==2.32.3
PyYAML==6.0.2
